made it work

+++ InversionCounter.cpp
@@ -23,34 +23,35 @@
 class InversionCounter {
 public: 
     // vector passed by ref! Will be mutated.  
-// Sorts an entire array and returns its total number of inversions.
-    int SortAndCountInversions(std::vector<int>& vector) {
+    // Sorts an entire array and returns its total number of inversions.
+    unsigned long SortAndCountInversions(std::vector<unsigned long>& vector) {
         // base case, array is already sorted and there are no inversions.
         if (vector.size() == 1)
             return 0;
 
         // Split array in half using iterators.
         std::size_t const halfSize = vector.size() / 2;
-        std::vector<int> lowerSplit(vector.begin(), vector.begin() + halfSize);
-        std::vector<int> upperSplit(vector.begin() + halfSize, vector.end());
+        std::vector<unsigned long> lowerSplit(vector.begin(), vector.begin() + halfSize);
+        std::vector<unsigned long> upperSplit(vector.begin() + halfSize, vector.end());
 
         // recursively sort array.
-        int lowerInversions = SortAndCountInversions(lowerSplit);
-        int upperInversions = SortAndCountInversions(upperSplit);
+        unsigned long lowerInversions = SortAndCountInversions(lowerSplit);
+        unsigned long upperInversions = SortAndCountInversions(upperSplit);
 
         // total inversions is sum of low, hi and split.
         MergeAndCountSolution sol = MergeAndCountInversions(lowerSplit, upperSplit);
         // we can do better here rather than returning a new vector.
         vector = sol.sortedVector;
+        
         return lowerInversions + upperInversions + sol.vectorplitInversions;
     }
 
     // Go through each value in the array and check how many values after the current one 
     // are less than the current one.
-    int BruteForceInvs(std::vector<int>& vector) {
-        int invs = 0;
-        for (int i = 0; i < vector.size() - 1; ++i) {
-            for (int j = i + 1; j < vector.size(); ++j) {
+    unsigned long BruteForceInvs(std::vector<unsigned long>& vector) {
+        unsigned long invs = 0;
+        for (unsigned long i = 0; i < vector.size() - 1; ++i) {
+            for (unsigned long j = i + 1; j < vector.size(); ++j) {
                 if (vector[j] < vector[i])
                     ++invs;
             }
@@ -58,15 +59,15 @@ public:
         return invs;
     }
 
-    std::vector<int> parseTextFile(const std::string& nameTextFile){
+    std::vector<unsigned long> parseTextFile(const std::string& nameTextFile){
         std::ifstream stream { nameTextFile };
 
         if (!stream) {
             throw "Nooooooo";
         }
 
-        std::vector<int> returnVec;
-        int row;
+        std::vector<unsigned long> returnVec;
+        unsigned long row;
         while (stream >> row) {
             returnVec.push_back(row);
 
@@ -78,21 +79,21 @@ public:
 private: 
     // This could apply to both solutions really, or can use references to reduce instantiated vectors. 
     static struct MergeAndCountSolution {
-        std::vector<int> sortedVector;
-        int vectorplitInversions;
+        std::vector<unsigned long> sortedVector;
+        unsigned long vectorplitInversions;
     };
 
     // returns num of split inversions, note that vector is passed by ref! Will be mutated. 
     // Merges and sorts two given arrays, assuming those arrays are sorted. Also returns number of split inversions.
-    MergeAndCountSolution MergeAndCountInversions(std::vector<int>& lowerSplit, std::vector<int>& upperSplit) {
+    MergeAndCountSolution MergeAndCountInversions(std::vector<unsigned long>& lowerSplit, std::vector<unsigned long>& upperSplit) {
 
         // divide and conquer, merge sort!
 
         MergeAndCountSolution sol;
-        int lowerCounter = 0; int upperCounter = 0; int splitInv = 0;
+        unsigned long lowerCounter = 0; unsigned long upperCounter = 0; unsigned long splitInv = 0;
 
         // Do we need parathesese around the sizes here?
-        for (int k = 0; k < lowerSplit.size() + upperSplit.size(); ++k) {
+        for (unsigned long k = 0; k < lowerSplit.size() + upperSplit.size(); ++k) {
             // check if lower vec has already been fully iterated through.
             if (lowerCounter == lowerSplit.size()) {
                 sol.sortedVector.push_back(upperSplit[upperCounter]);
